+++ lib/shared/src/cody-ignore/ignore-helper.ts
@@ -75,10 +75,12 @@ export class IgnoreHelper {
 
             // Build the ignore rule with the relative folder path applied to the start of each rule.
             for (let ignoreLine of ignoreFile.content.split('\n')) {
-                // Skip blanks/ comments
-                ignoreLine = ignoreLine.trim()
+                // Trim off any trailing comments.
+                ignoreLine = ignoreLine.split('#')[0]
 
-                if (!ignoreLine.length || ignoreLine.startsWith('#')) {
+                // Skip any lines that are now empty.
+                ignoreLine = ignoreLine.trim()
+                if (!ignoreLine.length) {
                     continue
                 }