+++ vscode/src/edit/utils/edit-models.ts
@@ -1,0 +1,32 @@
+import { ModelProvider } from '@sourcegraph/cody-shared'
+import type { AuthProvider } from '../../services/AuthProvider'
+import { type EditModel, ModelUsage } from '@sourcegraph/cody-shared/src/models/types'
+import type { EditIntent } from '../types'
+
+export function getEditModelsForUser(authProvider: AuthProvider): ModelProvider[] {
+    const authStatus = authProvider.getAuthStatus()
+    if (authStatus?.configOverwrites?.chatModel) {
+        ModelProvider.add(
+            new ModelProvider(authStatus.configOverwrites.chatModel, [
+                ModelUsage.Chat,
+                // TODO: Add configOverwrites.editModel for separate edit support
+                ModelUsage.Edit,
+            ])
+        )
+    }
+    return ModelProvider.get(ModelUsage.Edit, authStatus.endpoint)
+}
+
+export function getOverridenModelForIntent(intent: EditIntent, currentModel: EditModel): EditModel {
+    switch (intent) {
+        case 'doc':
+        case 'fix':
+        case 'test':
+            // Edit commands have only been tested with Claude 2. Default to that for now.
+            return 'anthropic/claude-2.0'
+        case 'add':
+        case 'edit':
+            // Support all model usage for add and edit intents.
+            return currentModel
+    }
+}