+++ vscode/src/non-stop/FixupController.ts
@@ -5,6 +5,7 @@ import {
     type ChatEventSource,
     type ContextFile,
     type ContextMessage,
+    type EditModel,
 } from '@sourcegraph/cody-shared'
 
 import { executeEdit } from '../edit/execute'
@@ -26,8 +27,8 @@ import { FixupScheduler } from './FixupScheduler'
 import { FixupTask, type taskID } from './FixupTask'
 import type { FixupFileCollection, FixupIdleTaskRunner, FixupTextChanged } from './roles'
 import { CodyTaskState } from './utils'
-import type { EditSupportedModels } from '../edit/prompt'
 import { getInput } from '../edit/input/get-input'
+import type { AuthProvider } from '../services/AuthProvider'
 
 // This class acts as the factory for Fixup Tasks and handles communication between the Tree View and editor
 export class FixupController
@@ -44,7 +45,7 @@ export class FixupController
 
     private _disposables: vscode.Disposable[] = []
 
-    constructor() {
+    constructor(private readonly authProvider: AuthProvider) {
         // Register commands
         this._disposables.push(
             vscode.workspace.registerTextDocumentContentProvider('cody-fixup', this.contentStore),
@@ -163,13 +164,14 @@ export class FixupController
         range: vscode.Range,
         expandedRange: vscode.Range | undefined,
         mode: EditMode,
-        model: EditSupportedModels,
+        model: EditModel,
         intent: EditIntent,
         contextMessages: ContextMessage[],
         source: ChatEventSource
     ): Promise<FixupTask | null> {
         const input = await getInput(
             document,
+            this.authProvider,
             {
                 initialRange: range,
                 initialExpandedRange: expandedRange,
@@ -207,7 +209,7 @@ export class FixupController
         selectionRange: vscode.Range,
         intent: EditIntent,
         mode: EditMode,
-        model: EditSupportedModels,
+        model: EditModel,
         source?: ChatEventSource,
         contextMessages?: ContextMessage[],
         destinationFile?: vscode.Uri
@@ -1072,6 +1074,7 @@ export class FixupController
         // Prompt the user for a new instruction, and create a new fixup
         const input = await getInput(
             document,
+            this.authProvider,
             {
                 initialInputValue: task.instruction,
                 initialRange: task.selectionRange,